Refactor to support other chat protocols
Moved the interfaces to the gobot package so we can implement them in places
other than just the xmpp bot.  Moved the client code to cmd/gobot/gobot.go so the
'gobot' package just contains the interfaces for now, and the implementation
lives in cmd.

Moved the Plugin interface from gobotplugings package to the main gobot package.