Fix imports that got munged while fixing merge conflicts
There was also an issue with XMPP using Reply(), it doesn't currently
work since it tries to reply to a user instead of a room.  I changed the
XMPP Reply() function to do the same thing as Send() for now, until we
have time to debug multi-room support for XMPP.  Slack multi-room still
works as expected.